991118-1.c: Cast various constants back to long long.
        * gcc.c-torture/execute/991118-1.c: Cast various constants back
        to long long.
        * gcc.c-torture/execute/bf64-1.c: Likewise.

From-SVN: r49336

+++ gcc/testsuite/gcc.c-torture/execute/991118-1.c
@@ -57,23 +57,22 @@ struct tmp4 tmp4 = {0x1FFFF00000000LL, 0x123};
 
 main()
 {
-
   if (sizeof (long long) != 8)
     exit (0);
 
   tmp = sub (tmp);
   tmp2 = sub2 (tmp2);
 
-  if (tmp.pad != 0x123 || tmp.field != 0xFFF9551175BDFDB5LL)
+  if (tmp.pad != 0x123 || tmp.field != (long long) 0xFFF9551175BDFDB5LL)
     abort ();
-  if (tmp2.pad != 0x123 || tmp2.field != 0xFFF9551175BDFDB5LL)
+  if (tmp2.pad != 0x123 || tmp2.field != (long long) 0xFFF9551175BDFDB5LL)
     abort ();
 
   tmp3 = sub3 (tmp3);
   tmp4 = sub4 (tmp4);
-  if (tmp3.pad != 0x123 || tmp3.field != 0xFFF989AB12345678LL)
+  if (tmp3.pad != 0x123 || tmp3.field != (long long) 0xFFF989AB12345678LL)
     abort ();
-  if (tmp4.pad != 0x123 || tmp4.field != 0xFFF989AB12345678LL)
+  if (tmp4.pad != 0x123 || tmp4.field != (long long) 0xFFF989AB12345678LL)
     abort ();
   exit (0);
 }
